Refactor views

+++ babble/views/follower_views.py
@@ -1,6 +1,7 @@
 from typing import List, Optional, Type
 
 from django.db import DatabaseError, transaction
+from django.db.models import F
 from django.db.models.manager import BaseManager
 from django.http import HttpRequest
 from rest_framework import status, viewsets
@@ -24,18 +25,20 @@ class FollowerViewSet(viewsets.ModelViewSet):
         try:
             with transaction.atomic():
                 serializer.save()
+
                 user: Optional[User] = User.objects.get_or_none(
                     pk=request.data.get("user")
                 )
                 following: Optional[User] = User.objects.get_or_none(
                     pk=request.data.get("following")
                 )
+
                 if user is None or following is None:
                     raise DatabaseError
-                user.following += 1
-                following.followers += 1
-                user.save()
-                following.save()
+
+                user.update(following=F("following") + 1)
+                following.update(followers=F("followers") + 1)
+
         except DatabaseError:
             return Response(
                 {"error": "Cancle follower"}, status=status.HTTP_409_CONFLICT
@@ -46,26 +49,21 @@ class FollowerViewSet(viewsets.ModelViewSet):
             status=status.HTTP_201_CREATED,
         )
 
-    def destroy(self, request, pk: Optional[int] = None) -> Response:
+    def destroy(self, request: HttpRequest, pk: Optional[int] = None) -> Response:
         following: Optional[User] = User.objects.get_or_none(pk=pk)
-        if following is None:
-            return Response(
-                {"error": "Following not found"}, status=status.HTTP_404_NOT_FOUND
-            )
         follower: Optional[Follower] = Follower.objects.get_or_none(
             User=request.user, following=following
         )
-        if follower is None:
-            return Response(
-                {"error": "Follower not found"}, status=status.HTTP_404_NOT_FOUND
-            )
+
+        if following is None or follower is None:
+            return Response({"error": "Not Found"}, status=status.HTTP_404_NOT_FOUND)
+
         try:
             with transaction.atomic():
+                request.user.update(following=F("following") - 1)
+                following.update(followers=F("followers") - 1)
                 follower.delete()
-                request.user.following -= 1
-                following.followers -= 1
-                request.user.save()
-                following.save()
+
         except DatabaseError:
             return Response(
                 {"error": "Cancle follower"}, status=status.HTTP_409_CONFLICT
@@ -79,10 +77,12 @@ class FollowerViewSet(viewsets.ModelViewSet):
     def get_followings(self, request: HttpRequest) -> Response:
         follower: List[Follower] = self.queryset.filter(follower=request.user)
         serializer: FollowerSerializer = FollowerSerializer(follower, many=True)
+
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     @action(detail=False, methods=["get"], url_path="followers")
     def get_followers(self, request: HttpRequest) -> Response:
         follower: List[Follower] = self.queryset.filter(following=request.user)
         serializer: FollowerSerializer = FollowerSerializer(follower, many=True)
+
         return Response(serializer.data, status=status.HTTP_200_OK)
